+++ src/DocumentSearch.php
@@ -16,6 +16,7 @@ use craft\console\Application as ConsoleApplication;
 use craft\elements\Asset;
 use craft\events\DefineBehaviorsEvent;
 use craft\events\RegisterElementSearchableAttributesEvent;
+use craft\helpers\App;
 use venveo\documentsearch\behaviors\AssetContentBehavior;
 use venveo\documentsearch\models\Settings;
 use venveo\documentsearch\services\DocumentContentService;
@@ -34,10 +35,6 @@ use yii\base\Event;
  */
 class DocumentSearch extends Plugin
 {
-    /**
-     * @var DocumentSearch
-     */
-    public static $plugin;
 
     /**
      * @inheritdoc
@@ -45,7 +42,6 @@ class DocumentSearch extends Plugin
     public function init()
     {
         parent::init();
-        self::$plugin = $this;
 
         if (Craft::$app instanceof ConsoleApplication) {
             $this->controllerNamespace = 'venveo\documentsearch\console\controllers';
@@ -70,7 +66,7 @@ class DocumentSearch extends Plugin
     /**
      * @inheritdoc
      */
-    protected function createSettingsModel()
+    protected function createSettingsModel(): ?\craft\base\Model
     {
         return new Settings();
     }
@@ -78,14 +74,14 @@ class DocumentSearch extends Plugin
     /**
      * @inheritdoc
      */
-    protected function settingsHtml(): string
+    protected function settingsHtml(): ?string
     {
         return Craft::$app->view->renderTemplate(
             'document-search/settings',
             [
                 'settings' => $this->getSettings(),
                 'binaries' => [
-                    'pdftotext' => is_file(Craft::parseEnv($this->getSettings()->pdfToTextExecutable))
+                    'pdftotext' => is_file(App::parseEnv($this->getSettings()->pdfToTextExecutable))
                 ]
             ]
         );