+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Document Search Changelog
 
+## 4.0.0 - 2022/07/24
+
+### Added
+- Added support for MS Office Documents. (.doc, .docx, .xlsx, .pptx, .txt)
+
+### Changed
+- Volume settings now store the volume uuid instead of ID - **ensure sure you update your settings**
+- Document Search now requires Craft 4
+
 ## 1.0.3 - 2020-10-27
 ### Fixed
 - Fix environment variable for pdftotext path not working (#5)